implemented Services , Controller Successfully

+++ HealthTrakerAPI/Dtos/HealthDataDto.cs
@@ -6,7 +6,7 @@
         public int UserId { get; set; }
         public int Steps { get; set; }
         public int HeartRate { get; set; }
-        public float SleepDuration { get; set; }
+        public double SleepDuration { get; set; }
         public DateTime Date { get; set; }
     }
 }
