+++ HealthTrakerAPI/Dtos/UserDto.cs
@@ -5,7 +5,7 @@
         public int UserId { get; set; }
         public string Name { get; set; }
         public int Age { get; set; }
-        public float Weight { get; set; }
-        public float Height { get; set; }
+        public double Weight { get; set; }
+        public double Height { get; set; }
     }
 }
