+++ HealthTrakerAPI/Models/User.cs
@@ -5,8 +5,8 @@
         public int UserId { get; set; }
         public string Name { get; set; }
         public int Age { get; set; }
-        public float Weight { get; set; }
-        public float Height { get; set; }
+        public double Weight { get; set; }
+        public double Height { get; set; }
         public ICollection<HealthData> HealthData { get; set; }
     }
 }
